+++ rules/array/tgt_unsafe.rs
@@ -11,10 +11,8 @@ unsafe fn f1<T1>(a0: Vec<T1>) -> *const T1 {
     a0.as_ptr()
 }
 
-// TODO: this should return usize. However std::size_t is translated as unsigned long which in turn
-// gets translated to u64.
-unsafe fn f2<T1>(a0: Vec<T1>) -> u64 {
-    a0.len() as u64
+unsafe fn f2<T1>(a0: Vec<T1>) -> usize {
+    a0.len()
 }
 
 unsafe fn f3<T1>(a0: &mut Vec<T1>) -> *mut T1 {
